Moving the encode middleware functions to the postLogin controller
