+++ controllers/login.js
@@ -1,23 +1,43 @@
 const createHttpError = require('http-errors')
+const bcryptjs = require('bcryptjs')
+
+const { User } = require('../database/models')
+
+const validationDb = require('../helpers/validationDb')
 const { endpointResponse } = require('../helpers/success')
 const { catchAsync } = require('../helpers/catchAsync')
+const { encode } = require('../helpers/jwtFuntions')
+const { ErrorObject } = require('../helpers/error')
 
 const postLogin = catchAsync(async (req, res, next) => {
-    try {
-        const response = { token: req.token }
+  const { email, password } = req.body
+  try {
+    const schema = { where: { email } }
+    const user = await validationDb(schema, User, true)
 
-        endpointResponse({
-            res,
-            message: 'Post Login',
-            body: response
-        })
-    } catch (error) {
-        const httpError = createHttpError(
-            error.statusCode,
-            `[User Not Found] - [index - POST]: ${error.message}`
-        )
-        next(httpError)
+    const compare = bcryptjs.compareSync(password, user.password)
+
+    if (!compare) {
+      const httpError = new ErrorObject('Invalid User or Password', 404)
+      next(httpError)
     }
+
+    const payload = { id: user.id }
+    const token = await encode(payload)
+    const response = { token }
+
+    endpointResponse({
+      res,
+      message: 'Post Login',
+      body: response
+    })
+  } catch (error) {
+    const httpError = createHttpError(
+      error.statusCode,
+            `[User Not Found] - [index - POST]: ${error.message}`
+    )
+    next(httpError)
+  }
 })
 
 module.exports = postLogin
